Use an Api class for abstraction
It is currently very basic as functions will be added to it when the time
comes. The idea is to have a method to, for example, show notifications
(and store them in a list with metadata) to the user or also to easily
show a message box overlaying the current display.

Private class members will permit encapsulations of variables related
to these features and limit the privilage that each app has on the
device.