Fix race condition on HOME press while rendering
This existed because waiting for the app to finish rendering was done in
a blocking manner, meaning that the rendering state would never end.
This would block the microcontroller in a loop.

This was fixed by rewriting the part related to app switching to
correctly do the switching while the app is not rendering.