Implement an app_manager that supports BaseApp
and create classes for home_menu and main_clock apps.

This commit has a lot of changes as multiple parts of the project had to be
changed to support running apps that are based on the BaseApp class. It
could not have been done in multiple commits  as the progam would not be
able to build and thus I would not be able to test the changes.