Update license to only GNU AGPL
I though about it more and it would be simpler to only have the GNU AGPL
license. Changing license shouldn't cause any problems as I am the only
one who has worked on this project. I'll keep the previous commit as it
shows the evolution of the project's licensing, even if brief.